requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
llama-index
llama-parser
openai
langchain